minor improvements to wsdl-schema gen
