+++ modules/openapi-generator/src/main/resources/wsdl-schema/jaxb-customization.mustache
