+++ modules/openapi-generator/src/main/resources/wsdl-schema/wsdl-converter.mustache
