+++ samples/schema/petstore/wsdl-schema/.openapi-generator/FILES
