+++ samples/schema/petstore/wsdl-schema/jaxb-customization.xml
