+++ samples/schema/petstore/wsdl-schema/service.wsdl
